requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.4
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
outcome==1.1.0
psycopg2-binary==2.8.6
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
